backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
click==8.2.1
fastapi==0.116.1
filelock==3.19.1
fsspec==2025.9.0
h11==0.16.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
setuptools==80.9.0
sniffio==1.3.1
starlette==0.47.3
sympy==1.14.0
torch==2.8.0
torchvision==0.23.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
